pyproject.toml
[project]
name = "rills"
version = "0.1.0"
description = "A Mafia game with LLM-powered characters"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
